refactor: extract remote-target resolution and provenance into target.py
Moves 15 top-level symbols (10 functions/helpers + 5 constants) from
server.py into a new leaf module, src/comfy_mcp/target.py, byte-for-byte:
everything about pointing the run/job tools at a ComfyUI running
ELSEWHERE via COMFYUI_URL/COMFYUI_HOST(+PORT) instead of the implicit
local default — resolution and redaction (_comfy_target,
_redact_config_url, _strip_brackets, _redact_target_host,
_format_target_endpoint, _malformed_target_note), the --host/--port
forward (_with_target, _TARGET_AWARE_SUBCOMMANDS), the local-only
download_model refusal (_remote_shared_models_optin,
_reject_remote_model_download, REMOTE_SHARED_MODELS_ENV,
_REMOTE_SHARED_MODELS_TRUE), DEFAULT_COMFYUI_PORT, and the
system_stats/free_memory divergence notes (_COMFY_TARGET_NOTE_KEY,
_annotate_comfy_target, _target_provenance_suffix, _comfy_cli_ran,
_with_target_provenance). Three contiguous blocks moved verbatim
(constants 100-202, resolution/redaction functions 749-1045,
annotation/provenance functions 5367-5561 in pre-move server.py);
verified byte-for-byte present in target.py via a source-substring
check against pre-move HEAD.

server.py reaches every moved name module-qualified (target._x) at 16
call sites and prose cross-references, via a word-boundary rename
(ruff F821 came back clean). server_info had a local variable also
named `target` (the resolved (host, port, source) tuple) that the
blanket rename would have silently shadowed the module import inside
that one function — same class of collision the argv/errors
extractions hit — so the LOCAL was renamed to `resolved_target`
instead, disclosed here rather than left implicit. textutil dropped
from server.py's imports (its only user, _redact_config_url, moved).

tests/conftest.py's autouse _clear_comfyui_target_env fixture renames
server.REMOTE_SHARED_MODELS_ENV -> target.REMOTE_SHARED_MODELS_ENV.
tests/test_remote_host.py renames 33 read-only references
(server._comfy_target 22, server.DEFAULT_COMFYUI_PORT 5,
server.REMOTE_SHARED_MODELS_ENV 6) to target.*; prose-only mentions
(docstrings/comments with no `server.` prefix) are left unqualified,
matching the clitext._x-without-import precedent. Zero monkeypatches
touch these names, so no test-fixture rewiring beyond the rename.

One pre-existing test went stale from the move and needed a real fix,
not just a rename: test_address_env_docs.py's
test_server_reads_the_comfyui_vars_and_not_comfy_local_url asserted
that COMFYUI_URL/HOST/PORT are read specifically inside server.py's
own source text. That reader is now target.py, so the assertion is
generalized to "read by ANY module in the package" — the same
ownership question the test's own COMFY_LOCAL_URL negative check
already asks across every source file, just applied to the positive
side too.

AGENTS.md's module table gains a target.py row ("Seven leaf modules"
-> "Eight").

Full suite: 1889 passed, 3 deselected. ruff check and ruff format
--check both clean.
